iterate RIF comms tests
